+++ test/int/test_http_server.py
@@ -1,6 +1,7 @@
 from conftest import Cluster, Instance
 from urllib.request import urlopen
 import pytest
+import json
 
 
 @pytest.fixture
@@ -37,3 +38,37 @@ def test_webui(cluster_with_webui: Cluster):
     instance.wait_online()
     with urlopen(f"http://{http_listen}/") as response:
         assert response.headers.get("content-type") == "text/html"
+
+    with urlopen(f"http://{http_listen}/api/v1/replicaset") as response:
+        assert response.headers.get("content-type") == "application/json"
+        assert json.load(response) == [
+            {
+                "grade": "Online",
+                "version": "??.??",
+                "instances": [
+                    {
+                        "failureDomain": {},
+                        "isLeader": True,
+                        "currentGrade": "Online",
+                        "targetGrade": "Online",
+                        "name": "i1",
+                        "version": "??.??",
+                    }
+                ],
+                "instanceCount": 1,
+                "capacity": 100,
+                "uuid": instance.replicaset_uuid(),
+                "id": "r1",
+            }
+        ]
+
+    with urlopen(f"http://{http_listen}/api/v1/cluster") as response:
+        assert response.headers.get("content-type") == "application/json"
+        assert json.load(response) == {
+            "capacityUsage": 100,
+            "replicasetsCount": 1,
+            "instancesCurrentGradeOffline": 0,
+            "currentInstaceVersion": "23.06.0",
+            "memory": {"usable": 33554432, "used": 33554432},
+            "instancesCurrentGradeOnline": 1,
+        }